+++ zetaclient/chains/evm/observer/observer_test.go
@@ -69,11 +69,9 @@ func getAppContext(
 
 	// feed chain params
 	err := appContext.Update(
-		observertypes.Keygen{},
 		[]chains.Chain{evmChain, chains.ZetaChainMainnet},
 		nil,
 		chainParams,
-		"tssPubKey",
 		*sample.CrosschainFlags(),
 	)
 	require.NoError(t, err)
@@ -117,7 +115,7 @@ func MockEVMObserver(
 	}
 	// use default mock tss if not provided
 	if tss == nil {
-		tss = mocks.NewTSSMainnet()
+		tss = mocks.NewTSS(t).FakePubKey(testutils.TSSPubKeyMainnet)
 	}
 	// create AppContext
 	appContext, _ := getAppContext(t, chain, "", &params)
@@ -182,7 +180,7 @@ func Test_NewObserver(t *testing.T) {
 			chainParams: params,
 			evmClient:   evmClient,
 			evmJSONRPC:  mocks.NewMockJSONRPCClient(),
-			tss:         mocks.NewTSSMainnet(),
+			tss:         mocks.NewTSS(t),
 			logger:      base.Logger{},
 			ts:          nil,
 			fail:        false,
@@ -200,7 +198,7 @@ func Test_NewObserver(t *testing.T) {
 				return evmClient
 			}(),
 			evmJSONRPC: mocks.NewMockJSONRPCClient(),
-			tss:        mocks.NewTSSMainnet(),
+			tss:        mocks.NewTSS(t),
 			logger:     base.Logger{},
 			ts:         nil,
 			fail:       true,
@@ -214,7 +212,7 @@ func Test_NewObserver(t *testing.T) {
 			chainParams: params,
 			evmClient:   evmClient,
 			evmJSONRPC:  mocks.NewMockJSONRPCClient(),
-			tss:         mocks.NewTSSMainnet(),
+			tss:         mocks.NewTSS(t),
 			before: func() {
 				envVar := base.EnvVarLatestBlockByChain(chain)
 				os.Setenv(envVar, "invalid")