+++ changelog.md
@@ -2,11 +2,13 @@
 
 ## Unreleased
 
-### Features
-
 ### Tests
 
-### Refactor
+* [3205](https://github.com/zeta-chain/node/issues/3205) - move Bitcoin revert address test to advanced group to avoid upgrade test failure
+
+## Refactor
+
+* [3170](https://github.com/zeta-chain/node/pull/3170) - revamp TSS package in zetaclient
 
 ### Fixes